# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateMainPageEntryRequest:
    properties:
      icon_url:
        type: string
      position:
        minimum: 1
        type: integer
      project_id:
        type: integer
    required:
    - icon_url
    - position
    - project_id
    type: object
...
  model.MainPageProject:
    properties:
      icon_url:
        type: string
      id:
        type: integer
      position:
        type: integer
      project:
        allOf:
        - $ref: '#/definitions/model.Project'
        description: Вграждаме целия съществуващ модел Project
    type: object
...
  model.Project:
    properties:
      age_categories:
        items:
          $ref: '#/definitions/model.Age_Category'
        type: array
      categories:
        items:
          $ref: '#/definitions/model.Category'
        type: array
      cover_image_url:
        example: https://example.com/poster.png
        type: string
      created_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      description:
        example: A short description of the project.
        type: string
      director:
        example: Jane Doe
        type: string
      duration:
        example: 120
        type: integer
      episodes:
        items:
          $ref: '#/definitions/model.Episode'
        type: array
      genres:
        items:
          $ref: '#/definitions/model.Genre'
        type: array
      id:
        example: 1
        type: integer
      is_featured:
        example: false
        type: boolean
      keywords:
        example: action,thriller
        type: string
      producer:
        example: John Smith
        type: string
      release_year:
        example: 2024
        type: integer
      screenshots:
        items:
          $ref: '#/definitions/model.ProjectScreenshot'
        type: array
      seasons:
        items:
          $ref: '#/definitions/model.Season'
        type: array
      title:
        example: Example Movie
        type: string
      type:
        example: movie
        type: string
    type: object
...
  model.UpdateMainPageEntryRequest:
    properties:
      icon_url:
        type: string
      position:
        minimum: 1
        type: integer
    required:
    - icon_url
    - position
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Server for Ozinse Video Platform Backend
  title: Ozinse API
  version: "1.0"
paths:
...
  /api/main-page-entries:
    get:
      consumes:
      - application/json
      description: Returns all projects assigned to the featured main page layout.
        Requires administrator role.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.MainPageProject'
            type: array
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List main page entries (Admin only)
      tags:
      - main-page
    post:
      consumes:
      - application/json
      description: Shortcuts a project onto the featured landing main page area. Requires
        administrator role.
      parameters:
      - description: Main page entry payload
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/model.CreateMainPageEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Main page entry created successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add project to main page (Admin only)
      tags:
      - main-page
  /api/main-page-entries/{id}:
    delete:
      consumes:
      - application/json
      description: Removes a featured shortcut entry from the main page by its config
        ID. Requires administrator role.
      parameters:
      - description: Main Page Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Main page entry deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete main page entry (Admin only)
      tags:
      - main-page
    get:
      consumes:
      - application/json
      description: Returns details of a main page featured layout configuration by
        its database ID. Requires administrator role.
      parameters:
      - description: Main Page Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.MainPageProject'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get main page entry by ID (Admin only)
      tags:
      - main-page
    put:
      consumes:
      - application/json
      description: Modifies position or meta layout details of an entry on the main
        page by its ID. Requires administrator role.
      parameters:
      - description: Main Page Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Main page entry values
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/model.UpdateMainPageEntryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Main page entry updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update main page entry (Admin only)
      tags:
      - main-page
  /api/projects:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of projects along with metadata
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Search projects by title
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Includes 'data' array and 'pagination' object
          schema:
            type: object
        "400":
          description: Invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List projects with pagination
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Creates a new project. Requires administrator role.
      parameters:
      - description: Project payload
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/model.CreateProjectDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create project (Admin only)
      tags:
      - projects
  /api/projects/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a project by ID. Requires administrator role.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: successfully deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete project (Admin only)
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Returns a project by its ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Project not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Updates an existing project by ID. Requires administrator role.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project payload
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/model.CreateProjectDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin only
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update project (Admin only)
      tags:
      - projects
  /api/projects/{id}/similar:
    get:
      consumes:
      - application/json
      description: Returns projects that are related to the given project ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Project'
            type: array
        "400":
          description: Invalid project ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get similar projects
      tags:
      - projects
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"